Revert "add support for http method HEAD (#599)" (#603)
This reverts commit 1b3bcbde7bedcb7213ca7657549867720c7eaf3e.

HEAD semantics are more complicated, per James:
We probably want to revert this. The specified contract of HEAD is to return the headers that would be returned if the resource was requested with a GET. Given that Conjure offers no normal support for reading out response headers, it feels a little dubious to let people override?

For example, if I do a HEAD and it sends me back Content-Length: 50, what that means is that if I do a GET, I should get back something of content length 50 - Conjure's generated server interfaces won't guarantee this properly.

Indeed, even Jersey will have a bad time - my understanding is that both OkHttp and Jersey will discard a response body provided for a HEAD request, so the test code is non-functional.