Fix references with container prefixes `set`/`list`/`optional`
Previously importing conjure objects using a reference along the
lines of `settings.Foo` would fail because it would match the
`set` prefix. Now we handle it correctly.

This introduces a change to the way parser markers work:
Instead of maintaining a mark/rewind/release stack, the mark
method returns an object which can be used to return to the
marked location, and the `release` functionality is no longer
necessarily. This takes advantage of the application stack
rather than maintaining a separate stack that can get out of
sync.